requirements.txt
google-adk
fastmcp
google-adk[a2a]
requests
numpy
pandas
openmeteo-requests
dotenv